requirements.txt
mozilla-django-oidc==1.2.2
PyJWT==1.7.1
requests==2.22.0
